+++ __mocks__/rimraf.js
@@ -1,0 +1,9 @@
+/* eslint-env jest */
+
+module.exports = jest.fn((p, options, cb) => {
+  if (typeof options === 'function') {
+    options();
+  } else {
+    cb();
+  }
+});