+++ src/__tests__/outdated.test.js
@@ -26,8 +26,6 @@ const spyFinalSpec = (spy, pkgName) => {
 };
 
 describe('OUTDATED command', () => {
-  beforeEach(() => { jest.resetAllMocks(); });
-
   it('does not modify any package.json', async () => {
     const writeSpecs = require('../utils/writeSpecs').default;
     const base = path.join(process.cwd(), 'test/fixtures/packages');