+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+* Add **`oao clean`** to remove all `node_modules` directories in sub-packages.
 * Extract Parallel Console (now published as `storyboard-listener-console-parallel` under the Storyboard monorepo).
 
 ## 0.7.0 (Feb. 27, 2017)